inital attemt at a body copy component that uses some serif fonts available on the system
