+++ src/@primitives/typography/index.js
@@ -5,4 +5,4 @@ export { default as H4 } from './H4';
 export { default as H5 } from './H5';
 export { default as H6 } from './H6';
 export { default as H7 } from './H7';
-export { default as P } from './P';
+export { default as BodyCopy } from './BodyCopy';
